% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_fcn.R
\name{poly_fcn}
\alias{poly_fcn}
\title{Apply a polynomial function to a number}
\usage{
poly_fcn(x)
}
\arguments{
\item{x}{A number or vector of numbers}
}
\value{
A number or vector of numbers resulting from applying the function to x.
}
\description{
This function applies the polynomial equation (x+1)^2 + (x+2)^3 to a number,
and produces an error message if one of the inputs is not numeric.
}
\examples{
poly_fcn(5)
poly_fcn(1:5)
}
