% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynom.R
\name{polynom.double}
\alias{polynom.double}
\title{Make polynom out of a vector}
\usage{
\method{polynom}{double}(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a language object
}
\description{
Makes a polynom out of vector x.
}
