% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynom.R
\name{polynom.list}
\alias{polynom.list}
\title{Construct from list}
\usage{
\method{polynom}{list}(x)
}
\arguments{
\item{x}{a list where the names are the powers and values are the coeficcients}
}
\description{
Construct from list
}
