% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynom.R
\name{polynom}
\alias{polynom}
\title{Create polynom}
\usage{
polynom(x, ...)
}
\arguments{
\item{x}{an argument vector or list}

\item{...}{write like "'3' = 2", and so on like in polynom.list}
}
\value{
Returns a language object
}
\description{
Create polynom
}
