% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynom.R
\name{polynom.vector}
\alias{polynom.vector}
\title{Make polynom out of a vector}
\usage{
\method{polynom}{vector}(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a language objec
}
\description{
Makes a polynom out of vector x.
}
