% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynom.R
\name{polynom.default}
\alias{polynom.default}
\title{Make a polynom}
\usage{
\method{polynom}{default}(x = NULL, ...)
}
\arguments{
\item{x}{an argument vector}

\item{...}{write like "'3' = 2", and so on like in polynom.list}
}
\description{
Make a polynom
}
