\name{zSummary.polynomial}
\alias{Summary.polynomial}
\title{Summary Group Methods for Polynomials}
\description{
  Allows summary functions to be used on polynomial arguments.
}
\usage{
Summary.polynomial(x, \dots)
}
\arguments{
  \item{x}{an object of class \code{"polynomial"}.}
  \item{\dots}{optional additional arguments.}
}
\details{
  Summary functions on polynomials do not make very much sense, so an
  error is returned.
}
\seealso{
  \code{\link{Math.polynomial}},
  \code{\link{Ops.polynomial}}
}
\keyword{symbolmath}
